vulnerability refactored and tested
